requirements.txt
pypdf2
python-dateutil
gunicorn
flask_wtf
requests
flask
werkzeug
pysendpulse
pymongo
rq
dnspython
mongo-queue-service
